Move styles to styles.js
