+++ src/views/Candidates/Candidates.jsx
@@ -1,18 +1,7 @@
 import React from 'react';
 import { Typography, Grid } from '@material-ui/core';
 import ApplicationCard from '../../components/ApplicationCard';
-import { makeStyles } from '@material-ui/core/styles';
-
-const useStyles = makeStyles(({ spacing }) => ({
-  heading: {
-    fontSize: spacing(8),
-  },
-  container: {
-    backgroundColor: '#F3F4F7',
-    padding: '10px',
-    height: '100vh',
-  },
-}));
+import useStyles from './styles';
 
 const candidateList = [
   {
